Exercise1
I used python function method to come up with the answer.

First I use the def even_numbers to define the method I will be using
Second step : I created a user input ’n’ to prompt the user
Because we are looking for an even_number I set the range from 2 and the user input +1 and also 2 to give  output of even numbers. I use letter ‘i’ to search through the range of the numbers and Its prints output after every input. And I call the even_number function ones again .